freertos-app: Adds freertos-app class and demo recipe
Adds a recipe to build a demo recipe for freertos.

It uses the freertos-app class to fetch the FreeRTOS source code
and the BSP repo (Linker Script and Makefile).

It uses these two in conjunction to build an example app
that works on top of FreeRTOS.

This app can be emulated using QEMU simply by running runqemu

Signed-off-by: Alejandro Hernandez <aehs29@gmail.com>