freertos-image: Introduce a FreeRTOS image class and split the BSP functionality
The whole idea of this layer is to create an abstraction between
the FreeRTOS kernel, the BSP and the application code.

Previously the freertos-app class would be inherited from a
recipe, it would provide applications a BSP to be built for,
and the image wiring code needed for it to work consistently
with OE-Core.

For scalability purposes, if we were to create more FreeRTOS
ports, it makes sense to split the image and BSP concepts,
each of these on a different class, this way, a new BSP port
could reuse the same code required for the image, while still
using different portable code.

Signed-off-by: Alejandro Hernandez <aehs29@gmail.com>