meta-freertos: Add a multiconfig example that works out of the box
Create a multiconfig build example for users to see its
possible to build Linux and FreeRTOS at the same time.

A dummy x86-64 config is created that uses qemux86-64 as
MACHINE, and builds the poky DISTRO.

This does not affect the layer normal behavior in which
running:

$ bitbake freertos-demo

Would still only build the FreeRTOS sample application.

But, running:

$ bitbake mc:dummy-x86-64:core-image-minimal

Would build both Linux and FreeRTOS at the same time.

To test Linux using QEMU, using the runqemu script and
passing the qemuboot.conf file for the respective build
works seamlessly.

To test FreeRTOS no argument to runqemu is required since
the FreeRTOS build is considered the default config.

e.g.

$ runqemu tmp-qemux86-64-glibc/deploy/images/qemux86-64/core-image-minimal-qemux86-64.qemuboot.conf

or

$ runqemu

Signed-off-by: Alejandro Hernandez <aehs29@gmail.com>