freertos_echo: Create a test case to use along with -c testimage
Create a test case with a known behavior that allows us to check
if the RTOS was built and is being executed properly.

Invoking the following:

$ bitbake freertos-demo-local -c testimage

Runs the RTOS on QEMU and automatically runs the test case mentioned above,
which checks for a certain output to be found on the serial console, if the
output is found it means the RTOS is working properly, if it is not found,
or a timeout is reached, then throw an error.

The message below is a sample of what the test report should look like:

RESULTS:
RESULTS - freertos_echo.FreeRTOSTest.test_freertos_echo: PASSED (2.00s)
SUMMARY:
freertos-demo-local () - Ran 1 test in 2.006s
freertos-demo-local - OK - All required tests passed (successes=1, skipped=0, failures=0, errors=0)

Signed-off-by: Alejandro Hernandez <aehs29@gmail.com>
Signed-off-by: Alejandro Hernandez Samaniego <alejandro@enedino.org>