layer.conf: Disable multiconfig setup by default
Initially we wanted to showcase the usage of multiconfig builds
by building an RTOS along with Linux at the same time, this has
been tested and proven to be useful, however, it seems reasonable
to believe that users would like to avoid this behavior by default.

Comment out the BBMULTICONFIG variable along with the dependency
to it, this will allow users to use this layer without having to
parse a separate multiconfig, such confs will still be in the layer
for anyone to use by simply uncommenting the BBMULTICONFIG
definition.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandro@enedino.org>