freertos-image.bbclass: Fix TESTIMAGE_BOOT_PATTERNS[search_reached_prompt]
The search_reached_prompt pattern implies the system is ready
to receive an input, before we used the FreeRTOS keyword, which
did confirm the device had booted but at that point it still
was not ready to receive an input from the keyboard.

Use Blocked... as a replacement pattern, since the RTOS booted,
and the task running became blocked it is ready to receive an
input to unblock such task.

This fixes a case on which automated testing failed since the
system wasnt able to detec the user had "logged in".

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandro@enedino.org>