meta-freertos: Enable STM32 build
Create new MACHINE=stm32f446 that allow builds for an ARM Cortex-M4F architecture.

An example application is provided to be used with the above MACHINE,
bitbake freertos-demo-stm32 can be invoked to build such example that
includes a bundled FreeRTOS kernel, the application can be then flashed
to the device using the .elf or .hex file located on the DEPLOY directory.

- Enable a CI build for this architecture, artifacts,
can be downloaded and flashed directly without the need to build manually.
- Modify the README to reflect on the latest changes.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandro@enedino.org>