meta-freertos: Refactor classes
From experience, weve learned that while it is possible to keep the Application,
BSP and Kernel source codes in separate repositores its not as useful as
initially thought due to the fact that for RTOS applications the BSP and
Application code is intertwined with one another. It is simpler to keep these
sources together as one repository.

Refactor classes, removing the BSP class, and passing the BSP related code that
is still required to the recipe instead, e.g. PORT, QB_foo variables.

- Introduce the FREERTOS_BUNDLE variable, used to specify whether the
application repository is bundled with the FreeRTOS kernel or not.

Set FREERTOS_BUNDLE=1 if the application code already contains a FreeRTOS kernel
Set FREERTOS_BUNDLE=0 if we want to use the FreeRTOS kernel code provided by the
class/layer. 0 is default.

The FreeRTOS included kernel is conditionally included only when
FREERTOS_BUNDLE=0.

- Fix optimization flags
- Use FREERTOS_SRC_DIR and FREERTOS_KERNEL_SRC_DIR to point to where the
FreeRTOS sources are located.
- Conditionally set the FREERTOS_VERSION variable since it can now be set by the
recipe in case it contains a bundled FreeRTOS kernel.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandro@enedino.org>